🔧 重構通知系統模組架構 - 解決循環依賴問題
🎯 核心修復:
• 將所有類型定義移至 notifications.types.ts 作為單一來源
• 移除 notificationsApi.ts 和 notifications.types.ts 之間的循環依賴
• 修復模組解析問題：'requested module does not provide export'

📁 架構改進:
• notifications.types.ts: 集中定義所有類型、介面和列舉
• notificationsApi.ts: 從類型檔案導入，專注於API邏輯
• 避免重新導出(re-export)造成的模組解析問題

✅ 問題解決:
• SyntaxError: requested module does not provide export 'Notification'
• 模組循環依賴導致的編譯錯誤
• Vite HMR 模組緩存問題

🚀 技術改進:
• 清晰的依賴方向：types <- services <- hooks <- components
• 單一來源真實性：所有通知相關類型統一管理
• 避免ES模組重新導出的陷阱

🧪 測試狀態:
• React 開發服務器編譯成功，無錯誤
• 模組導入解析正常
• 準備進行瀏覽器端測試驗證
